🤖 Refactor(client): header, api요청 리팩토링
