+++ apps/client/package.json
@@ -20,6 +20,7 @@
         "nanoid": "^5.0.8",
         "react": "^18.3.1",
         "react-dom": "^18.3.1",
+        "react-router-dom": "^7.0.1",
         "react-syntax-highlighter": "^15.6.1",
         "react-type-animation": "^3.2.0",
         "terraform": "file:../../packages/terraform"
