+++ apps/client/src/Root.tsx
@@ -1,0 +1,15 @@
+import { useLoaderData } from 'react-router-dom';
+import CloudGraphProvider from '@components/CloudGraphProvider';
+import Layout from '@components/Layout';
+
+function Root() {
+    const loader = useLoaderData();
+
+    return (
+        <CloudGraphProvider initialData={loader?.data.architecture}>
+            <Layout />
+        </CloudGraphProvider>
+    );
+}
+
+export default Root;
