+++ apps/client/src/apis/index.ts
@@ -1,12 +1,13 @@
+const BASE_URL = import.meta.env.VITE_API_URL;
 export const URLS = {
     login: 'auth/login',
     privateArchi: (id: string) => `private-architectures/${id}`,
 };
-export const urls = (domain: string, path: keyof typeof URLS, slug?: any) => {
+export const urls = (path: keyof typeof URLS, slug?: any) => {
     const urls = URLS[path];
     if (typeof urls === 'function') {
-        return `${domain}/${urls(slug)}`;
+        return `${BASE_URL}/${urls(slug)}`;
     }
 
-    return `${domain}/${urls}`;
+    return `${BASE_URL}/${urls}`;
 };
